+++ src/game/logic/game.loader.ts
@@ -83,6 +83,8 @@ export class GameLoader implements OnModuleInit {
 
   private async loadTrainer(region: Region, area: AreaDocument, object: any, map: any) {
     const starters = getProperty<string>(object, 'Starters');
+    const monsterSpecs = JSON.parse(getProperty<string>(object, 'Monsters') || '[]');
+
     const trainer = await this.trainerService.upsert({
       region: region._id.toString(),
       area: area._id.toString(),
@@ -103,16 +105,15 @@ export class GameLoader implements OnModuleInit {
       y: (object.y / map.tileheight) | 0,
       direction: getProperty<number>(object, 'Direction') ?? Direction.DOWN,
       'npc.encounterOnSight': getProperty<boolean>(object, 'EncounterOnSight') || false,
+      'npc.encounterOnTalk': monsterSpecs?.length > 0,
       'npc.canHeal': getProperty<boolean>(object, 'CanHeal') || false,
       'npc.walkRandomly': getProperty<boolean>(object, 'WalkRandomly') || false,
       'npc.path': getProperty<string>(object, 'Path')?.split(/[,;]/g)?.map(s => +s) || null,
       'npc.starters': starters ? JSON.parse(starters) : undefined,
     });
 
-    const monsterSpecs = JSON.parse(getProperty<string>(object, 'Monsters') || '[]');
     trainer.team = [];
-    for (const monsterSpec of monsterSpecs) {
-      const [type, level] = monsterSpec;
+    for (const [type, level] of monsterSpecs) {
       const monster = await this.monsterGeneratorService.createAuto(trainer._id.toString(), type, level);
       trainer.team.push(monster._id.toString());
     }