+++ src/game/logic/npc-talk/npc-talk.service.ts
@@ -47,15 +47,10 @@ export class NpcTalkService {
           $addToSet: {'npc.encountered': trainerId},
         });
         await this.monsterGeneratorService.createAuto(trainerId, starterId, 1);
-        await this.trainerService.update(dto._id, {
-          $addToSet: {
-            encounteredMonsterTypes: starterId,
-          },
-        });
       }
     }
-    if (target.npc.encounterOnSight) {
-      await this.trainerService.update(dto._id, {
+    if (target.npc.encounterOnTalk) {
+      await this.trainerService.update(target_id, {
         $addToSet: {'npc.encountered': trainerId},
       });
       await this.battleSetupService.createTrainerBattle(target, [trainer]);