+++ docs/Changelog.md
@@ -109,6 +109,41 @@
 * Fixed the `GET /groups` endpoints causing an `500 Internal Server Error` when the `members` query parameter was present multiple times as an array.
 * Fixed a potential problem where newly created monsters would not trigger `created` events.
 
+# v2.1.3 - UDP Fixes
+
+## Improvements
+
+* UDP commands are now properly validated and send an `error` event when invalid.
+
+## Bugfixes
+
+* Fixed a problem with UDP sessions that would not unsubscribe correctly.
+* It is no longer possible to teleport to other areas using trainer move commands.
+
+# v2.1.4 - Presets Rate Limits
+
+## General
+
++ Added rate limits for presets (some endpoints use different rate limits).
+
+# v2.1.5 - Trainer Cleanup
+
+## Improvements
+
+* Trainers without any progress near spawn are now deleted after a while.
+
+## Bugfixes
+
+* Fixed an edge case where Roof tiles would allow walking through walls.
+* Fixed some missing parameters in Swagger.
+
+# v2.1.6 - Fix Ghost Trainers
+
+## Bugfixes
+
+* Fixed a bug where trainers would still block tiles after being deleted (manually or during cleanup).
+* Preset requests for `@2x` images now return `404 Not Found` instead of the original image.
+
 # v3.0.0 - Battle
 
 ## New Features
@@ -116,24 +151,28 @@
 ### Content
 
 + Added the Encounter Test region.
++ Added Route 101, 102 and 111.
++ Added Adromere.
++ Overhauled Route 100.
++ Optimized all maps and tilesets -- some maps now use layer data instead of chunks.
 
 ### Resources
 
 + Added Encounters.
-+ Added Opponents
-  > This is the representation of a trainer in battle.
-+ Added `TalkTrainerDto`
-  > This UDP action can be used to talk to trainers.
-+ Added `encounterOnSight`, `canHeal`, `encountered` and `starters` to `Trainer.npc`.
++ Added Opponents, the representation of a trainer in battle.
++ Added `TalkTrainerDto`, a UDP action for talking to trainers.
++ Added `encounterOnSight`, `encounterOnTalk`, `canHeal`, `encountered` and `starters` to `NPCInfo`.
 + Added `team` to `Trainer`.
 + Added `encounteredMonsterTypes` to `Trainer`.
-  > I.e. the Mondex.
 
 ## Improvements
 
 * Trainers no longer receive a monster when they are created.
-  > Visit the Prof. to select your start monster.
+* Players can receive a starter monster from the Prof.
 * Wild Encounters can be triggered by Tall Grass.
 * Players can trigger encounters by talking to other trainers.
 * Players can heal their monsters by talking to a Nurse.
 * NPCs may start encounters when they see you.
+* Defeating a monster now grants experience.
+* Monsters can now level up and evolve.
+* Winning encounters now grants coins.